Move dumpToFile(ByteBuffer) function to SystemHelpers
